Refactor lib/utils/parse dependency in api package.
