+++ api/types/traits.go
@@ -16,14 +16,6 @@ limitations under the License.
 
 package types
 
-import (
-	"github.com/gravitational/teleport/lib/utils"
-	"github.com/gravitational/teleport/lib/utils/parse"
-
-	"github.com/gravitational/trace"
-	log "github.com/sirupsen/logrus"
-)
-
 // TraitMapping is a mapping that maps a trait to one or
 // more teleport roles.
 type TraitMapping struct {
@@ -37,84 +29,3 @@ type TraitMapping struct {
 
 // TraitMappingSet is a set of trait mappings
 type TraitMappingSet []TraitMapping
-
-// TraitsToRoles maps the supplied traits to a list of teleport role names.
-func (ms TraitMappingSet) TraitsToRoles(traits map[string][]string) []string {
-	var roles []string
-	ms.traitsToRoles(traits, func(role string, expanded bool) {
-		roles = append(roles, role)
-	})
-	return utils.Deduplicate(roles)
-}
-
-// TraitsToRoleMatchers maps the supplied traits to a list of role matchers. Prefer calling
-// this function directly rather than calling TraitsToRoles and then building matchers from
-// the resulting list since this function forces any roles which include substitutions to
-// be literal matchers.
-func (ms TraitMappingSet) TraitsToRoleMatchers(traits map[string][]string) ([]parse.Matcher, error) {
-	var matchers []parse.Matcher
-	var firstErr error
-	ms.traitsToRoles(traits, func(role string, expanded bool) {
-		if expanded || utils.ContainsExpansion(role) {
-			// mapping process included variable expansion; we therefore
-			// "escape" normal matcher syntax and look only for exact matches.
-			// (note: this isn't about combatting maliciously constructed traits,
-			// traits are from trusted identity sources, this is just
-			// about avoiding unnecessary footguns).
-			matchers = append(matchers, literalMatcher{
-				value: role,
-			})
-			return
-		}
-		m, err := parse.NewMatcher(role)
-		if err != nil {
-			if firstErr == nil {
-				firstErr = err
-			}
-			return
-		}
-		matchers = append(matchers, m)
-	})
-	if firstErr != nil {
-		return nil, trace.Wrap(firstErr)
-	}
-	return matchers, nil
-}
-
-// TraitsToRoles maps the supplied traits to teleport role names and passes them to a collector.
-func (ms TraitMappingSet) traitsToRoles(traits map[string][]string, collect func(role string, expanded bool)) {
-	for _, mapping := range ms {
-		for traitName, traitValues := range traits {
-			if traitName != mapping.Trait {
-				continue
-			}
-		TraitLoop:
-			for _, traitValue := range traitValues {
-				for _, role := range mapping.Roles {
-					outRole, err := utils.ReplaceRegexp(mapping.Value, role, traitValue)
-					switch {
-					case err != nil:
-						if trace.IsNotFound(err) {
-							log.WithError(err).Debugf("Failed to match expression %v, replace with: %v input: %v", mapping.Value, role, traitValue)
-						}
-						// this trait value clearly did not match, move on to another
-						continue TraitLoop
-						// skip empty replacement or empty role
-					case outRole == "":
-					case outRole != "":
-						collect(outRole, outRole != role)
-					}
-				}
-			}
-		}
-	}
-}
-
-// literalMatcher is used to "escape" values which are not allowed to
-// take advantage of normal matcher syntax by limiting them to only
-// literal matches.
-type literalMatcher struct {
-	value string
-}
-
-func (m literalMatcher) Match(in string) bool { return m.value == in }
